Time: 138 ms (77.57%), Space: 82.2 MB (50.48%) - LeetHub

+++ 0853-car-fleet/0853-car-fleet.cpp
@@ -16,9 +16,9 @@ public:
         
         stack<float> st;
         for(int i = 0; i < n; i++) {
-            int pos = car[i].first;
-            int speed = car[i].second;
-            float time = (float)(target-pos)/speed;
+            float pos = car[i].first;
+            float speed = car[i].second;
+            float time = (target-pos)/speed;
             
             while(!st.empty() && st.top() <= time) {
                 st.pop();
